style startup screen
