+++ app/src/main/res/drawable/ic_contentful_logo.xml
@@ -1,0 +1,51 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:height="177dp"
+        android:viewportHeight="177.1625"
+        android:viewportWidth="867.0875"
+        android:width="867dp"
+        >
+    <path
+            android:fillAlpha="1"
+            android:fillColor="#f3e705"
+            android:fillType="nonZero"
+            android:pathData="m53.506,123.656 l0.019,-0.019c-8.975,-8.988 -14.538,-21.388 -14.538,-35.05 0,-13.688 5.563,-26.087 14.55,-35.063l-0.025,-0.025c7.612,-7.613 7.612,-19.95 0,-27.55 -7.613,-7.613 -19.95,-7.613 -27.563,-0C9.913,41.974 0,64.124 0,88.586c0,24.462 9.913,46.6 25.95,62.625 7.612,7.612 19.95,7.612 27.563,-0 7.612,-7.6 7.612,-19.938 0,-27.55 -0.002,-0.002 -0.004,-0.004 -0.007,-0.006"
+            android:strokeColor="#00000000"
+            />
+    <path
+            android:fillAlpha="1"
+            android:fillColor="#65bbe7"
+            android:fillType="nonZero"
+            android:pathData="m53.512,53.503 l0.019,0.019c8.987,-8.975 21.388,-14.537 35.05,-14.537 13.688,-0 26.088,5.563 35.063,14.55l0.025,-0.025c7.612,7.612 19.95,7.612 27.55,-0 7.613,-7.613 7.613,-19.95 0,-27.563 -16.025,-16.038 -38.175,-25.95 -62.638,-25.95 -24.462,-0 -46.599,9.913 -62.625,25.95 -7.612,7.613 -7.612,19.95 0,27.563 7.6,7.612 19.938,7.612 27.55,-0 0.002,-0.002 0.003,-0.004 0.006,-0.007"
+            android:strokeColor="#00000000"
+            />
+    <path
+            android:fillAlpha="1"
+            android:fillColor="#f03e48"
+            android:fillType="nonZero"
+            android:pathData="m123.663,123.657 l-0.019,-0.019C114.656,132.613 102.256,138.175 88.594,138.175c-13.688,-0 -26.087,-5.563 -35.063,-14.55l-0.025,0.025c-7.613,-7.612 -19.95,-7.612 -27.55,-0C18.344,131.263 18.344,143.6 25.956,151.212 41.981,167.25 64.131,177.163 88.594,177.163c24.462,-0 46.6,-9.913 62.625,-25.95 7.613,-7.612 7.613,-19.95 0,-27.562 -7.6,-7.612 -19.938,-7.612 -27.55,-0 -0.002,0.002 -0.004,0.004 -0.006,0.006"
+            android:strokeColor="#00000000"
+            />
+    <path
+            android:fillAlpha="1"
+            android:fillColor="#3781be"
+            android:fillType="nonZero"
+            android:pathData="m53.513,53.499c-7.613,7.612 -19.95,7.612 -27.563,-0 -7.613,-7.613 -7.613,-19.95 0,-27.55 7.612,-7.613 19.95,-7.613 27.563,-0 7.612,7.6 7.612,19.938 0,27.55"
+            android:strokeColor="#00000000"
+            />
+    <path
+            android:fillAlpha="1"
+            android:fillColor="#d03634"
+            android:fillType="nonZero"
+            android:pathData="m53.513,151.211c-7.613,7.612 -19.95,7.612 -27.563,-0 -7.613,-7.6 -7.613,-19.938 0,-27.55 7.612,-7.613 19.95,-7.613 27.563,-0C61.125,131.274 61.125,143.611 53.513,151.211"
+            android:strokeColor="#00000000"
+            />
+    <path
+            tools:ignore="VectorPath"
+            android:fillAlpha="1"
+            android:fillColor="#282b34"
+            android:fillType="evenOdd"
+            android:pathData="m500.88,78.699 l41.519,-0c-1.58,-10.329 -10.145,-14.417 -20.007,-14.417 -10.01,-0 -18.466,4.657 -21.511,14.417zM298.023,65.562c12.584,-0 21.049,10.981 21.049,22.98 0,12.656 -7.766,22.81 -21.049,22.81 -13.284,-0 -21.044,-10.158 -21.044,-22.81 0,-12.481 7.9,-22.98 21.044,-22.98l0,-0zM500.464,95.973c1.811,11.233 12.509,15.859 23.052,15.859 4.961,-0 10.693,-1.267 15.191,-3.39 0.771,-0.363 1.499,-0.687 2.236,-1.114 2.494,-1.44 4.781,-2.295 7.016,-0.011 1.205,1.232 2.411,2.457 3.639,3.669 1.161,1.152 2.209,2.302 3.225,3.581 3.38,4.262 1.009,5.271 -2.815,7.854 -8.208,5.547 -18.97,8.297 -28.816,8.297 -26.486,-0 -44.107,-16.021 -44.107,-42.821 0,-25.272 17.513,-42.019 42.664,-42.019 25.747,-0 40.456,15.117 40.923,40.816 0.126,7.032 0.44,9.278 -6.896,9.278l-55.311,-0 0,-0zM297.861,46.68c-24.384,-0 -41.533,17.699 -41.533,41.862 0,24.495 17.007,41.855 41.694,41.855 24.653,-0 41.862,-17.335 41.862,-41.855 0,-24.225 -17.722,-41.862 -42.024,-41.862l0,-0zM588.586,58.307c7.33,-6.959 15.269,-11.303 25.641,-11.303 22.446,-0 34.139,18.396 34.139,39.444l0,35.455c0,3.697 -3.029,6.724 -6.727,6.724l-7.201,-0c-3.697,-0 -6.726,-3.026 -6.726,-6.724l0,-35.296c0,-11.665 -6.28,-20.718 -18.63,-20.718 -12.06,-0 -19.759,10.224 -19.759,21.682l0,34.383c0,3.665 -3,6.673 -6.665,6.673l-7.15,-0c-3.668,-0 -6.674,-3.005 -6.674,-6.673l0,-67.153c0,-3.664 3.009,-6.668 6.674,-6.668l6.094,-0c3.721,-0 6.351,3.059 6.669,6.621l0.316,3.553 0,-0zM365.984,58.302c7.33,-6.957 15.273,-11.298 25.645,-11.298 22.442,-0 34.136,18.397 34.136,39.444l0,35.456c0,3.698 -3.031,6.722 -6.726,6.722l-7.205,-0c-3.696,-0 -6.721,-3.026 -6.721,-6.722l0,-35.297c0,-11.667 -6.284,-20.718 -18.635,-20.718 -12.059,-0 -19.753,10.229 -19.753,21.684l0,34.38c0,3.67 -3.006,6.673 -6.676,6.673l-7.145,-0c-3.675,-0 -6.674,-3.001 -6.674,-6.673 0,-21.414 0,-42.828 0,-64.241l0,-2.912c0,-3.669 3.004,-6.668 6.674,-6.668l6.093,-0c3.716,-0 6.356,3.064 6.674,6.624l0.314,3.546 0,-0zM456.166,48.45 L471.536,48.45c3.67,-0 6.669,3.007 6.669,6.677l0,4.568c0,3.67 -2.999,6.671 -6.669,6.671l-15.37,-0 0,33.756c0,6.493 3.115,10.245 9.701,10.245l3.12,-0c3.674,-0 6.682,2.989 6.682,6.671l0,4.569c0,3.674 -3.013,6.672 -6.682,6.672l-4.622,-0c-18.112,-0 -28.689,-7.754 -28.689,-28.157 0,-21.813 0,-43.625 0,-65.438 0,-3.671 3.009,-6.68 6.68,-6.68l7.139,-0c3.672,-0 6.671,3.009 6.671,6.68l0,13.766zM678.431,48.45 L693.799,48.45c3.67,-0 6.671,3.007 6.671,6.677l0,4.568c0,3.67 -3.001,6.671 -6.671,6.671l-15.368,-0 0,33.756c0,6.493 3.112,10.238 9.7,10.238l3.119,-0c3.676,-0 6.677,3.001 6.677,6.678l0,4.569c0,3.672 -3.007,6.672 -6.677,6.672l-5.935,-0c-1.129,0.026 -2.175,-0.019 -3.294,-0.07 -16.638,-1.023 -24.083,-11.81 -24.083,-28.087 0,-21.813 0,-43.625 0,-65.438 0,-3.671 3.004,-6.675 6.674,-6.675l7.144,-0c3.676,-0 6.675,3.004 6.675,6.675l0,13.766zM728.618,48.498 L742.666,48.498c3.713,-0 6.675,3.106 6.675,6.793l0,4.658c0,3.686 -2.96,6.792 -6.675,6.792l-14.049,-0 0,55.165c0,3.703 -3.024,6.722 -6.726,6.722l-7.2,-0c-3.697,-0 -6.729,-3.024 -6.729,-6.722l0,-76.212c0,-16.666 9.35,-29.763 26.362,-31.358 0.466,-0.043 0.925,-0.072 1.38,-0.094 0.999,-0.093 2.029,-0.065 3.015,-0.035l6.559,-0c3.714,-0 6.673,3.11 6.673,6.796l0,4.231c0,3.686 -2.956,6.791 -6.673,6.791l-4.806,-0.005c-0.161,0.024 -0.308,0.029 -0.469,0.029 -8.04,-0 -11.386,6.216 -11.386,13.644l0,2.805zM815.201,118.596c-7.805,7.446 -15.342,11.321 -26.46,11.321 -22.538,-0 -33.33,-18.621 -33.33,-39.45l0,-35.448c0,-3.697 3.025,-6.731 6.722,-6.731l7.206,-0c3.697,-0 6.726,3.034 6.726,6.731l0,35.293c0,11.662 6.27,20.718 18.62,20.718 12.086,-0 19.764,-10.261 19.764,-21.766l0.009,-12.675c-0.116,-0.496 -0.17,-0.976 -0.17,-1.486l0,-20.084c0,-3.699 3.026,-6.731 6.726,-6.731l7.195,-0c3.701,-0 6.731,3.029 6.731,6.731l0,66.934c0,3.67 -3.007,6.673 -6.676,6.673l-6.09,-0c-3.723,-0 -6.36,-3.062 -6.675,-6.628l-0.299,-3.402 0,-0zM853.156,128.626c-3.695,-0 -6.726,-3.024 -6.726,-6.722l0,-87.274c0,-3.695 3.031,-6.726 6.726,-6.726l7.202,-0c3.696,-0 6.726,3.029 6.726,6.726l0,87.274c0,3.698 -3.033,6.722 -6.726,6.722l-7.202,-0 0,-0zM238.258,106.924c5.571,-4.323 6.375,-1.844 11.158,2.881 2.426,2.392 8.903,6.476 4.616,10.021 -8.835,7.305 -16.782,10.622 -28.236,10.622 -25.655,-0 -43.814,-19.122 -42.451,-44.432 0.613,-11.398 5.026,-22.137 13.876,-29.592 8.125,-6.846 18.209,-9.881 28.772,-9.684 5.378,0.105 10.164,0.895 15.154,2.892 3.225,1.29 6.375,2.778 9.145,4.899 5.625,4.296 3.041,6.845 -1.359,11.247 -0.986,0.986 -1.935,2.006 -2.909,3.003 -2.191,2.239 -3.754,3.856 -6.787,1.78 -5.739,-3.926 -13.015,-6.039 -19.889,-4.259 -22.104,5.721 -20.925,45.093 6.125,45.093 4.497,-0 9.239,-1.719 12.785,-4.473l0,-0"
+            android:strokeColor="#00000000"
+            />
+</vector>
