--- app/src/main/res/layout/fragment_startup.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-             xmlns:tools="http://schemas.android.com/tools"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent"
-             tools:context=".home.StartupFragment"
-        >
-
-    <TextView
-            android:drawableTop="@mipmap/ic_launcher_foreground"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:layout_gravity="center"
-            android:text="@string/app_name"
-            android:scaleX="3"
-            android:scaleY="3"
-            android:textSize="@dimen/startup_app_text"
-            />
-
-</FrameLayout>
